@@ -16,21 +16,21 @@ project_urls =
 url = https://github.com/TranslationalML/pacsman
 
 [options]
-python_requires = >=3.8.17
+python_requires = >=3.10
 install_requires =  # file:requirements.txt
-    pandas  # >= 0.23.4
-    Cerberus >= 1.3.2
-    hypothesis >= 6.0.2
-    nibabel >= 2.3.0
-    nipype
-    pydicom >= 2.4.3
-    tqdm >= 4.43.0
-    beautifulsoup4 >= 4.12.2
+    pandas == 2.2
+    Cerberus == 1.3.5
+    hypothesis == 6.98
+    nibabel == 5.2
+    pydicom == 2.4
+    tqdm == 4.66
+    beautifulsoup4 == 4.12
+    jsonschema == 4.21
 test_requires =
-    pytest
-    pytest-console-scripts
-    pytest-cov
-    pytest-order >= 1.1.0
+    pytest == 7.4
+    pytest-console-scripts == 1.4
+    pytest-cov == 4.1
+    pytest-order == 1.2
 packages = find:
 
 [options.exclude_package_data]
@@ -38,29 +38,29 @@ packages = find:
 
 [options.extras_require]
 doc =
-    pydot >= 1.2.3
-    sphinx >= 7.1.2
-    sphinx-argparse == 0.4.0
-    sphinx_rtd_theme == 1.3.0
-    recommonmark == 0.7.1
-    sphinxcontrib-apidoc == 0.3.0
-    sphinxcontrib-mermaid == 0.9.2
-    sphinxemoji == 0.2.0
-    mock == 5.0.1
-    m2r2 == 0.3.2
-    docutils == 0.18.1
-    commonmark == 0.9.1
+    pydot == 2.0
+    sphinx == 7.2
+    sphinx-argparse == 0.4
+    sphinx_rtd_theme == 2.0
+    recommonmark == 0.7
+    sphinxcontrib-apidoc == 0.5
+    sphinxcontrib-mermaid == 0.9
+    sphinxemoji == 0.3
+    mock == 5.1
+    m2r2 == 0.3
+    docutils == 0.20
+    commonmark == 0.9
 dev =
-    black ~= 22.3.0
-    pre-commit
-    isort ~= 5.10.1
+    black == 24.2
+    pre-commit == 3.6
+    isort == 5.13
 docs =
     %(doc)s
 test =
-    pytest
-    pytest-cov
-    pytest-console-scripts
-    pytest-env
+    pytest == 7.4
+    pytest-console-scripts == 1.4
+    pytest-cov == 4.1
+    pytest-order == 1.2
 all =
     %(doc)s
     %(dev)s
@@ -91,13 +91,3 @@ ignore =
 per-file-ignores =
     **/__init__.py : F401
     docs/conf.py : E265
-
-[tool:pytest]
-norecursedirs = .git
-addopts = -svx
-doctest_optionflags = ALLOW_UNICODE NORMALIZE_WHITESPACE ELLIPSIS
-env =
-    PYTHONHASHSEED=0
-filterwarnings =
-    ignore::DeprecationWarning
-junit_family=xunit2
